Restructure the tokio shim
